Move logger module
